Update Env_parameters_correlation_RDA(4).R
Did not manage to follow the existing calculations of the RDA and etc.
Wrote new calculations of concentration for DCM and EPI, calculated RDA and ploted using  ggplot.

Did not go through the correlation tests. We should consider splitting the plot to physical and chemichal parameters, moreover we should introduce microscopy counts here!